Recognize CUDA_PATH environment variable
CUDA_PATH is the default used by CUDA v10.0 and v10.1, so we should recognize it in addition to CUDA_HOME.  Also find libraries in a CUDA_PATH/lib/ directory, as well as CUDA_PATH/lib64/ (lib/ is used by default in v10.0 and v10.1 on my 64-bit Linux and Windows platforms).  Lastly, check for nvcc.exe, since in Windows it has the file extension.